sidecar/manga-ocr/requirements.txt
manga-ocr>=0.1.14
Pillow>=10.0.0
